Creadas las entidades turno, ronda, partida, isla y las relaciones entre ellas
