--- src/main/java/org/springframework/samples/sevenislands/cardBaraja/CardBarajaService.java
@@ -1,26 +1,0 @@
-package org.springframework.samples.sevenislands.cardBaraja;
-
-import java.util.Collection;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CardBarajaService {
-    
-    @Autowired
-    private CardBarajaRepository cardRepository;
-
-    public Collection<CardBaraja> getCardsByType(CardPurpose cardPurpose){
-        return cardRepository.findByCardPurpose(cardPurpose);
-    }
-    
-    public Collection<CardType> getAllCardTypes(){
-        return cardRepository.findAllCardTypes();
-    }
-
-    public Collection<CardPurpose> getAllCardPurposes(){
-        return cardRepository.findAllCardPurposes();
-    }
-
-}
