--- src/main/java/org/springframework/samples/sevenislands/cardBaraja/CardType.java
@@ -1,26 +1,0 @@
-package org.springframework.samples.sevenislands.cardBaraja;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-import org.springframework.samples.sevenislands.model.BaseEntity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "CardTypes")
-public class CardType extends BaseEntity {
-    
-     @NotNull
-     @NotBlank
-     @Column(unique = true)
-     protected String name;
-
-
-}
