+++ src/main/java/org/springframework/samples/sevenislands/game/Game.java
@@ -9,6 +9,7 @@ import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.samples.sevenislands.lobby.Lobby;
@@ -29,10 +30,15 @@ public class Game extends BaseEntity {
 	@DateTimeFormat(pattern = "yyyy/MM/dd")
 	private LocalDateTime endingDate;
 
-	@OneToOne
+
+	@OneToOne(mappedBy = "game")
 	@NotNull
 	private Lobby lobby;
 
 	@OneToMany(cascade = CascadeType.ALL, mappedBy = "game", fetch = FetchType.LAZY)
 	private Set<Round> rounds;
+
+	@OneToMany(cascade = CascadeType.ALL, mappedBy = "game", fetch = FetchType.LAZY)
+	@Size(min = 7, max = 7)
+	private Set<Island> islands;
 }
