+++ src/main/java/org/springframework/samples/sevenislands/game/Island.java
@@ -3,12 +3,14 @@ package org.springframework.samples.sevenislands.game;
 import java.util.Set;
 
 import javax.persistence.Entity;
+import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+import org.springframework.samples.sevenislands.card.Card;
 import org.springframework.samples.sevenislands.model.BaseEntity;
 
 import lombok.Getter;
@@ -22,9 +24,8 @@ public class Island extends BaseEntity {
 
     @ManyToOne
     @NotNull
-    private Round round;
+    private Game game;
 
-    @OneToMany()
-    @Size(min = 1)
-    private Set<Turn> turns;
+    // @ManyToMany(mappedBy = "islands")
+    // private Set<Card> cards;
 }
