+++ src/main/java/org/springframework/samples/sevenislands/lobby/Lobby.java
@@ -7,8 +7,10 @@ import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
+import org.springframework.samples.sevenislands.game.Game;
 import org.springframework.samples.sevenislands.model.BaseEntity;
 import org.springframework.samples.sevenislands.player.Player;
 
@@ -30,6 +32,9 @@ public class Lobby extends BaseEntity{
 
     @OneToMany(cascade = CascadeType.PERSIST)
     private List<Player> players;
+
+    @OneToOne(cascade = CascadeType.ALL)
+    private Game game;
 
     public List<Player> getPlayerInternal() {
 		if (this.players == null) {
